endpoint: move functions to `labels.go`
Signed-off by: Ian Vernon <ian@cilium.io>
